+++ Demos/TextFormattingDemo.cs
@@ -15,7 +15,7 @@ namespace Demos {
 
         private const string Text =
             "MLEM's text formatting system allows for various <b>formatting codes</b> to be applied in the middle of a string. Here's a demonstration of some of them.\n\n" +
-            "You can write in <b>bold</i>, <i>italics</i>, <u>with an underline</u>, <st>strikethrough</st>, with a <s #000000 4>drop shadow</s> whose <s #ff0000 4>color</s> and <s #000000 10>offset</s> you can modify in each application of the code, or with various types of <b>combined <c Pink>formatting</c> codes</b>.\n\n" +
+            "You can write in <b>bold</i>, <i>italics</i>, <u>with an underline</u>, <st>strikethrough</st>, with a <s>drop shadow</s> whose <s #ff0000 4>color</s> and <s #000000 10>offset</s> you can modify in each application of the code, with an <o>outline</o> that you can also <o #ff0000 4>modify</o> <o #ff00ff 2>dynamically</o>, or with various types of <b>combined <c Pink>formatting</c> codes</b>.\n\n" +
             "You can apply <c CornflowerBlue>custom</c> <c Yellow>colors</c> to text, including all default <c Orange>MonoGame colors</c> and <c #aabb00>inline custom colors</c>.\n\n" +
             "You can also use animations like <a wobbly>a wobbly one</a>, as well as create custom ones using the <a wobbly>Code class</a>.\n\n" +
             "You can also display <i grass> icons in your text, and use super<sup>script</sup> or sub<sub>script</sub> formatting!\n\n" +
@@ -27,7 +27,13 @@ namespace Demos {
         private TokenizedString tokenizedText;
         private GenericFont font;
         private bool drawBounds;
-        private float scale = TextFormattingDemo.DefaultScale;
+        private float Scale {
+            get {
+                // calculate our scale based on how much larger the window is, so that the text scales with the window
+                var viewport = new Rectangle(0, 0, this.Game.Window.ClientBounds.Width, this.Game.Window.ClientBounds.Height);
+                return TextFormattingDemo.DefaultScale * Math.Min(viewport.Width / 1280F, viewport.Height / 720F);
+            }
+        }
 
         public TextFormattingDemo(MlemGame game) : base(game) {}
 
@@ -35,7 +41,10 @@ namespace Demos {
             this.Game.Window.ClientSizeChanged += this.OnResize;
 
             // creating a new text formatter as well as a generic font to draw with
-            this.formatter = new TextFormatter();
+            this.formatter = new TextFormatter {
+                DefaultShadowOffset = new Vector2(4),
+                DefaultOutlineThickness = 4
+            };
             // GenericFont and its subtypes are wrappers around various font classes, including SpriteFont, MonoGame.Extended's BitmapFont and FontStashSharp
             // supplying a bold and italic version of the font here allows for the bold and italic formatting codes to be used
             this.font = new GenericSpriteFont(
@@ -50,7 +59,7 @@ namespace Demos {
             // tokenizing our text and splitting it to fit the screen
             // we specify our text alignment here too, so that all data is cached correctly for display
             this.tokenizedText = this.formatter.Tokenize(this.font, TextFormattingDemo.Text, TextAlignment.Center);
-            this.tokenizedText.Split(this.font, this.GraphicsDevice.Viewport.Width * TextFormattingDemo.WidthMultiplier, this.scale, TextAlignment.Center);
+            this.tokenizedText.Split(this.font, this.GraphicsDevice.Viewport.Width * TextFormattingDemo.WidthMultiplier, this.Scale, TextAlignment.Center);
         }
 
         public override void DoDraw(GameTime time) {
@@ -59,7 +68,7 @@ namespace Demos {
 
             // we draw the tokenized text in the center of the screen
             // since the text is already center-aligned, we only need to align it on the y axis here
-            var size = this.tokenizedText.GetArea(Vector2.Zero, this.scale).Size;
+            var size = this.tokenizedText.GetArea(Vector2.Zero, this.Scale).Size;
             var pos = new Vector2(this.GraphicsDevice.Viewport.Width / 2, (this.GraphicsDevice.Viewport.Height - size.Y) / 2);
 
             // draw bounds, which can be toggled with B in this demo
@@ -67,13 +76,13 @@ namespace Demos {
                 var blank = this.SpriteBatch.GetBlankTexture();
                 this.SpriteBatch.Draw(blank, new RectangleF(pos - new Vector2(size.X / 2, 0), size), Color.Red * 0.25F);
                 foreach (var token in this.tokenizedText.Tokens) {
-                    foreach (var area in token.GetArea(pos, this.scale))
+                    foreach (var area in token.GetArea(pos, this.Scale))
                         this.SpriteBatch.Draw(blank, area, Color.Black * 0.25F);
                 }
             }
 
             // draw the text itself
-            this.tokenizedText.Draw(time, this.SpriteBatch, pos, this.font, Color.White, this.scale, 0);
+            this.tokenizedText.Draw(time, this.SpriteBatch, pos, this.font, Color.White, this.Scale, 0);
 
             this.SpriteBatch.End();
         }
@@ -91,13 +100,9 @@ namespace Demos {
         }
 
         private void OnResize(object sender, EventArgs e) {
-            // scale our text based on window size
-            var viewport = new Rectangle(0, 0, this.Game.Window.ClientBounds.Width,  this.Game.Window.ClientBounds.Height);
-            this.scale = TextFormattingDemo.DefaultScale * Math.Min(viewport.Width / 1280F, viewport.Height / 720F);
-
             // re-split our text if the window resizes, since it depends on the window size
             // this doesn't require re-tokenization of the text, since TokenizedString also stores the un-split version
-            this.tokenizedText.Split(this.font, this.GraphicsDevice.Viewport.Width * TextFormattingDemo.WidthMultiplier, this.scale, TextAlignment.Center);
+            this.tokenizedText.Split(this.font, this.GraphicsDevice.Viewport.Width * TextFormattingDemo.WidthMultiplier, this.Scale, TextAlignment.Center);
         }
 
     }