+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Jump to version:
 ## 6.2.0 (In Development)
 
 ### MLEM
+Additions
+- Added a simple outline formatting code
+
 Fixes
 - Fixed control characters being included in TextInput
 - Fixed TextInputs behaving incorrectly when switching between multiline and single-line modes